pyproject.toml
[project]
name = "label-red"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "opencv-python>=4.13.0.92",
    "pillow>=12.2.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0",
    "scipy>=1.17.1",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "transformers>=5.5.0",
    "uvicorn[standard]>=0.30.0",
]

[tool.uv.workspace]
members = [
    "PETA",
]
